dev_requirements.txt
lxml
pandas
pillow
matplotlib
python-dotenv
numpy
tensorflow
torchvision
google-cloud-storage
